Started the Camera class....
The goal of the camera class is to have a map bigger than the screen and
have the player always be at it's center when possible. The main goal
was to have the camera class take in actual x, y values and convert them
to render values based on platforms in range of the screen but physics
kinda died. Will investigate in the morning